pyproject.toml
[project]
name = "python-api-playground"
version = "0.1.0"
description = ""
authors = [
    {name = "aleksei.shemetov",email = "alxshemetov@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pytest (>=8.4.2,<9.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "allure-pytest (>=2.15.0,<3.0.0)",
    "pytest-xdist (>=3.8.0,<4.0.0)",
    "pydantic[email,timezone] (>=2.12.2,<3.0.0)",
    "datamodel-code-generator (>=0.35.0,<0.36.0)",
    "faker (>=37.11.0,<38.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "flask (>=3.1.2,<4.0.0)",
    "werkzeug (>=3.1.3,<4.0.0)",
    "pytest-check (>=2.6.0,<3.0.0)",
    "assertpy (>=1.1,<2.0)"
]

[tool.poetry]
packages = [{include = "python_api_playground", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
